refactor(admin): split ScenarioRunnerService into write/lifecycle/eval phase services (max-params 2/3)
ScenarioRunnerService had 5 injected deps (ingest, facts, entities, search,
surreal). Split per the "per-phase runners" brief into three focused services,
each ≤3 deps, plus an orchestrator that holds the run lifecycle:

- ScenarioWriteService (ingest) — additive ingest steps (fact/mention/link)
- ScenarioLifecycleService (facts, entities, surreal) — destructive steps
  (retract/forget), externalRef→entityId lookup, ephemeral-tenant teardown
- ScenarioEvalService (search) — query expectations (+debug trace), memory
  assertions, identity-merge verdict
- ScenarioRunnerService (write, lifecycle, evaluator) — setup-loop dispatch,
  metric aggregation, scenario catalogue (list/getById)

Result types moved to scenario-runner.types.ts (re-exported from the service
for baseline.service + admin controllers); pure helpers to
scenario-runner-utils.ts. Public API (list/getById/runOne/
cleanupEphemeralTenants) and behaviour unchanged.

tsc + typecheck + lint clean; 700 unit + 128 e2e + 9 jobs green.
